+++ test/egm2008.cpp
@@ -29,6 +29,8 @@ using namespace heyoka_test;
 TEST_CASE("mu/a getters")
 {
     REQUIRE(model::get_egm2008_a() != model::get_egm2008_mu());
+    REQUIRE(model::get_egm2008_CS().extent(0) > 0u);
+    REQUIRE(model::get_egm2008_CS().extent(1) == 2u);
 }
 
 TEST_CASE("error handling")
@@ -48,11 +50,10 @@ TEST_CASE("error handling")
                                    "geopotential model: the order 1 is greater than the degree 0"));
 
     REQUIRE_THROWS_MATCHES(
-        model::egm2008_acc({x, y, z}, model::detail::egm2008_max_degree, 0), std::invalid_argument,
-        Message(
-            fmt::format("Invalid degree specified for the computation of the acceleration in "
-                        "the EGM2008 geopotential model: a degree of {} is too high, the maximum allowed value is {}",
-                        model::detail::egm2008_max_degree, model::detail::egm2008_max_degree - 1u)));
+        model::egm2008_acc({x, y, z}, model::detail::egm2008_max_degree + 1u, 0), std::invalid_argument,
+        Message(fmt::format("Invalid degree specified for the EGM2008 geopotential model: the maximum degree is "
+                            "{}, but a degree of {} was specified",
+                            model::detail::egm2008_max_degree, model::detail::egm2008_max_degree + 1u)));
 }
 
 // A test to check that custom mu/a are correctly forwarded.
@@ -1252,7 +1253,7 @@ TEST_CASE("acceleration")
 
     // Run a computation up to the max degree/order.
     REQUIRE_NOTHROW(
-        model::egm2008_pot({x, y, z}, model::detail::egm2008_max_degree - 1u, model::detail::egm2008_max_degree - 1u));
+        model::egm2008_acc({x, y, z}, model::detail::egm2008_max_degree, model::detail::egm2008_max_degree));
 
     // Prepare input/output for the compiled functions.
     std::vector<double> out(3), in(3);
@@ -1275,7 +1276,7 @@ TEST_CASE("acceleration")
         REQUIRE(out[1] == approximately(test_accelerations[i][1]));
         REQUIRE(out[2] == approximately(test_accelerations[i][2]));
 
-        cf(out_grad, in);
+        cf_grad(out_grad, in);
         REQUIRE(out[0] == approximately(out_grad[0]));
         REQUIRE(out[1] == approximately(out_grad[1]));
         REQUIRE(out[2] == approximately(out_grad[2]));